refactor: infinite-carousel에서 performance-carousel로 네이밍 변경
