+++ packages/design-system/src/components/performance-carousel/hooks/use-carousel-data.ts
