+++ packages/design-system/src/components/performance-carousel/hooks/use-control-time.ts
