+++ packages/design-system/src/components/performance-carousel/hooks/use-data-format.ts
