+++ packages/design-system/src/components/performance-carousel/info-button/info-button.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 import InfoButton from './info-button';
 
 const meta: Meta<typeof InfoButton> = {
-  title: 'Common/InfiniteCarousel/InfoButton',
+  title: 'Common/PerformanceCarousel/InfoButton',
   component: InfoButton,
   parameters: {
     layout: 'centered',
