+++ packages/design-system/src/components/performance-carousel/mocks/bottom-banner-data.ts
