+++ packages/design-system/src/components/performance-carousel/performance-carousel.css.ts
