+++ packages/design-system/src/components/performance-carousel/progress-bar/progress-bar.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 import ProgressBar from './progress-bar';
 
 const meta: Meta<typeof ProgressBar> = {
-  title: 'Common/InfiniteCarousel/ProgressBar',
+  title: 'Common/PerformanceCarousel/ProgressBar',
   component: ProgressBar,
   parameters: {
     layout: 'centered',
