+++ packages/design-system/src/components/performance-carousel/progress-bar/progress-bar.tsx
